chore(heading): move to components/contentful folder
